# Faster Q3_K implementation on Metal