Similar to #2290 but for `Q3_K`, which is notoriously difficult to implement efficiently.

The table gives TG-128 time for `Q3_K_S` in ms/t on M2 Max with 30-core GPU:

| Model | Master | This PR | Speedup |
|--:|--:|--:|--:|
| 7B | 28.1 | 22.0 | 27.7% |
|13B | 50.3 | 37.5 | 34.1% |
|33B | 121.3 | 88.8 | 36.6% |
|65B | 230.3 | 167.8 | 37.2% |